create resources from the top-level accessURL and webService
But only use the top-level accessURL if there are no distributions, because if 'distribution' is used then the accessURL is supposed to be one of them (I think).

Write an extras field for which resource corresponds to the top-level accessURL so we can put the right one in generated data.json files when round-tripping.